+++ raise_utils/hyperparams/__init__.py
@@ -1,2 +1,3 @@
 from raise_utils.hyperparams.dodge import DODGE
+from raise_utils.hyperparams.ghost import GHOST
 from raise_utils.hyperparams.raytune import RayTune